ISP20200021 Shipping P03 1. 修正金額判斷時只取整數的問題

+++ Sci.Production.Shipping/P03.cs
@@ -248,7 +248,7 @@ where se.WKNo = '{0}' and se.junk=0", MyUtility.Convert.GetString(this.CurrentMa
 
             if (gridData.Rows.Count > 0 && this.chkImportChange.Checked)
             {
-                if (MyUtility.Convert.GetInt(this.CurrentMaintain["TPEPaidUSD"]) > 0)
+                if (MyUtility.Convert.GetDecimal(this.CurrentMaintain["TPEPaidUSD"]) > 0)
                 {
                     MyUtility.Msg.WarningBox("WK has been shared expense,  [No Import Charge] shouldn't tick, please double check.");
                 }
